Rename lazyfile file to file_abstraction
