+++ libimagstore/src/file_abstraction.rs
